requirements.txt
streamlit==1.32.0
pandas==2.1.4
langchain==0.1.13
langchain-openai==0.0.8
langchain-experimental==0.0.55
python-dotenv==1.0.0
openai==1.12.0